Adjust colors on load of SOE formatted map
When an SOE formatted map is loaded, automatically adjust colors in the
same way that colors were adjusted by the mapconvert script.  This
should make the SOE-formatted maps easier to see/read on SEQ's dark map
background.

Note:  the original colors are preserved if the map is saved back to the
SOE format after modification.